fix(*): code cleanups
