+++ wp-content/civi-extensions/goonjcustom/Civi/CollectionCampService.php
@@ -19,6 +19,7 @@ use Civi\Api4\Utils\CoreUtil;
 use Civi\Core\Service\AutoSubscriber;
 use Civi\Traits\CollectionSource;
 use Civi\Traits\QrCodeable;
+use Civi\Traits\SubtypeSource;
 
 /**
  *
@@ -26,6 +27,7 @@ use Civi\Traits\QrCodeable;
 class CollectionCampService extends AutoSubscriber {
   use QrCodeable;
   use CollectionSource;
+  use SubtypeSource;
 
   const FALLBACK_OFFICE_NAME = 'Delhi';
   const RELATIONSHIP_TYPE_NAME = 'Collection Camp Coordinator of';
@@ -565,8 +567,7 @@ class CollectionCampService extends AutoSubscriber {
    *   The parameters that were sent into the calling function.
    */
   public static function setOfficeDetails($op, $groupID, $entityID, &$params) {
-    $subtypeName = self::getSubtypeNameByEntityId($entityID);
-    if ($op !== 'create' || $subtypeName !== self::ENTITY_SUBTYPE_NAME) {
+    if ($op !== 'create' ||  self::getSubtypeNameByEntityId($entityID) !== self::ENTITY_SUBTYPE_NAME) {
       return;
     }
 
@@ -631,11 +632,11 @@ class CollectionCampService extends AutoSubscriber {
     else {
       $coordinator = $coordinators->first();
     }
-    
+
     if (!$coordinator) {
       throw new \Exception('No coordinator available to assign.');
     }
-    
+
     $coordinatorId = $coordinator['contact_id_a'];
 
     EckEntity::update('Collection_Camp', FALSE)
@@ -1141,23 +1142,6 @@ class CollectionCampService extends AutoSubscriber {
       ->addValue('Camp_Outcome.Number_of_Contributors', $contributorCount)
       ->addWhere('id', '=', $collectionCamp['id'])
       ->execute();
-  }
-
-  /**
-   *
-   */
-  public static function getSubtypeNameByEntityId($entityID) {
-    $getSubtypeName = civicrm_api4('Eck_Collection_Camp', 'get', [
-      'select' => [
-        'subtype:name',
-      ],
-      'where' => [
-            ['id', '=', $entityID],
-      ],
-    ]);
-    $entityData = $getSubtypeName[0] ?? [];
-    $subtypeName = $entityData['subtype:name'] ?? NULL;
-    return $subtypeName;
   }
 
 }
